Remove old download in favor of download-all
