+++ lib/commands/download.js
@@ -1,10 +1,10 @@
-const fs = require('fs-extra');
+const fs = require('fs');
 const path = require('path');
 const chalk = require('chalk');
-const unzip = require('unzipper');
-const api = require('@outdoorsyco/crowdin-api');
+const unzipper = require('unzipper');
 const loadConfig = require('../utils/config');
 const getClient = require('../utils/client');
+const api = require('@outdoorsyco/crowdin-api');
 const waitAndExport = require('../utils/wait-and-export');
 
 const CACHE_MINUTES = 5;
@@ -15,31 +15,34 @@ module.exports = {
   name: 'i18n:download',
   aliases: ['i18n:dl'],
   availableOptions: [
-    { name: 'assets', type: Boolean, default: false }
+    { name: 'assets', type: Boolean, default: true }
   ],
-  description: 'Download i18n files from Crowdin',
+  description: 'Download i18n files from Crowdin, including addons',
 
-  run: function(commandOptions) {
-    commandOptions = commandOptions || {};
-    const project = commandOptions.project || this.project;
-    const config = loadConfig(project.root);
+  run: function({
+    assets = true,
+    parentApp = this.project,
+    ui = this.ui
+  }) {
+    const includeAssets = assets;
+    this.parentApp = parentApp;
+    const addons = this._outdoorsyAddonsWithTranslations();
+    const config = loadConfig(parentApp.root);
     const crowdin = getClient(config);
-    const folderName = config.folderName || project.pkg.name;
-    const projectRoot = project.root;
 
-    this.ui.writeLine(chalk.cyan(`---- Getting translations for ${folderName} ----`));
+    ui.writeLine(chalk.cyan(`---- Getting all translations for ${parentApp.pkg.name} ----`));
 
     // First, check if the zip exists on disk already
     if (fs.existsSync(allTranslationsZip)) {
       const stats = fs.statSync(allTranslationsZip);
       const ageInMinutes = (new Date() - new Date(stats.mtime)) / 1000 / 60;
       if (ageInMinutes > CACHE_MINUTES) {
-        this.ui.writeLine(chalk.red('Translations cache expired'));
+        ui.writeLine(chalk.red('Translations cache expired'));
       } else {
-        this.ui.writeLine(chalk.green(
-          `Already have all translations downloaded. Getting ${folderName} files from disk`
+        ui.writeLine(chalk.green(
+          `Already have all translations downloaded. Getting files from disk`
         ));
-        return extractFiles({ folderName, projectRoot, commandOptions, config, ui: this.ui });
+        return extractFiles({ parentApp, addons, config, includeAssets, ui });
       }
     } else {
 
@@ -49,16 +52,16 @@ module.exports = {
 
     api.setKey(config.apiKey);
 
-    this.ui.writeLine(chalk.yellow(
+    ui.writeLine(chalk.yellow(
       `Building and downloading all assets and translations from crowdin`
     ));
-    return waitAndExport(config.projectName, folderName).then(result => {
-      this.ui.writeLine(`Crowdin build status: ${chalk.green(result.success.status)}`);
+    return waitAndExport(config.projectName).then(result => {
+      ui.writeLine(`Crowdin build status: ${chalk.green(result.success.status)}`);
       return new Promise((resolve, reject) => {
         // Download all translations and write zip to disk
         crowdin.download().pipe(fs.createWriteStream(allTranslationsZip))
         .on('finish', () => {
-          return extractFiles({ folderName, projectRoot, commandOptions, config, ui: this.ui }).then(() => {
+          return extractFiles({ parentApp, addons, config, includeAssets, ui }).then(() => {
             resolve();
           }).catch((error) => {
             reject(error);
@@ -69,47 +72,58 @@ module.exports = {
       throw error;
     });
   },
+
+  _outdoorsyAddonsWithTranslations() {
+    return this.parentApp.addons.filter((addon) => {
+      const path = addon.root;
+      return path.includes('outdoorsyco') && fs.existsSync(`${path}/config/crowdin.js`);
+    });
+  }
 };
 
 // Extract files for this folder from zip file to translation folder for correct project
-function extractFiles({ folderName, projectRoot, commandOptions, config, ui }) {
+function extractFiles({ parentApp, addons, config, includeAssets, ui }) {
   return new Promise((resolve, reject) => {
-    ui.writeLine(chalk.green(`Extracting ${folderName} assets and translations`));
+    ui.writeLine(chalk.green(`Extracting assets and translations`));
     fs.createReadStream(allTranslationsZip)
-    .pipe(unzip.Parse())
-    .on('entry', entry => {
-      let isThisProject = entry.path.split('/').includes(folderName);
-      if(isThisProject && entry.type !== 'Directory' && !entry.path.includes("en.yaml")){
-        let relPath = entry.path.replace(`${folderName}/`, '');
-        let fullPath = path.join(projectRoot, relPath);
-        let isAsset = !relPath.includes('.yaml');
-        let hasAssets = commandOptions && commandOptions.assets;
+    .pipe(unzipper.Parse())
+    .on('entry', file => {
+      // iterate through every translation file and put it in its correct place
+      const folderName = file.path.split('/')[0];
+      const matchedAppOrAddon = [parentApp].concat(addons).find((appOrAddon) => {
+        const appOrAddonCrowdinName = loadConfig(appOrAddon.root).folderName;
 
+        return folderName === appOrAddonCrowdinName;
+      })
+      if (file.type !== 'Directory' && matchedAppOrAddon && !file.path.includes("en.yaml")) {
+        const relativePath = file.path.replace(`${folderName}/`, '');
+        let fullPath = path.join(matchedAppOrAddon.root, relativePath);
+        const isAsset = !relativePath.includes('.yaml');
 
-        if(config.assetPaths && config.assetPaths.length && isAsset) {
-          if(hasAssets) {
+        if (config.assetPaths && config.assetPaths.length && isAsset) {
+          if (includeAssets) {
             config.assetPaths.forEach((asset) => {
-              let assetLocaleIndex = asset.exportPattern.split('/').indexOf('%locale%');
-              let splitPath = relPath.split('/');
-              let locale = splitPath[assetLocaleIndex];
+              const assetLocaleIndex = asset.exportPattern.split('/').indexOf('%locale%');
+              const splitPath = relativePath.split('/');
+              const locale = splitPath[assetLocaleIndex];
               splitPath[assetLocaleIndex] = locale.toLowerCase();
-              let lcLocalePath = path.join(...splitPath);
-              let relEntryPath = entry.path.replace(`${folderName}/`, '');
-              fullPath = path.join(projectRoot, lcLocalePath);
-              if(relEntryPath.toLowerCase() === lcLocalePath) {
-                createFolderFor(entry, fullPath);
+              const lcLocalePath = path.join(...splitPath);
+              const relativeFilePath = file.path.replace(`${folderName}/`, '');
+              fullPath = path.join(matchedAppOrAddon.root, lcLocalePath);
+              if(relativeFilePath.toLowerCase() === lcLocalePath) {
+                createFolderFor(file, fullPath);
               }
             });
           } else {
-            entry.autodrain();
+            file.autodrain();
             return;
           }
         }
 
-        entry.pipe(fs.createWriteStream(fullPath));
-        ui.writeLine(chalk.green(`Extracted file ${entry.path}`));
+        file.pipe(fs.createWriteStream(fullPath));
+        ui.writeLine(chalk.green(`Writing file ${file.path}`));
       } else {
-        entry.autodrain();
+        file.autodrain();
       }
     })
     .on('close', resolve)
