+++ lib/commands/validate.js
@@ -28,7 +28,7 @@ module.exports = {
       return this.ui.writeLine(chalk.green('Translations validated!'));
     } else {
       this.ui.writeLine(chalk.red('Some translations were missing.  Validation failed.'));
-      this.ui.writeLine(chalk.red('You may need to run ember i18n:download-all'));
+      this.ui.writeLine(chalk.red('You may need to run ember i18n:download'));
       throw('Failing build');
     }
